Makefile: 'distclean' improvement attempt 2
The 'distclean' target in automake-style projects means "clean everything". But
the $(BINDIST) target was assuming it to mean "clean everything that the
bindist target generates", which caused the Windows build to go wonky:

    http://ci.narc.ro/job/Cataclysm-Matrix/Graphics=Curses,Platform=Windows/4703/console

Create a new target 'clean-bindist' to only clean the bindist bits, and
repurpose 'distclean' to be the "clean everything" target.

Also update the .PHONY line to include 'clean' and 'clean-bindist'.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>